IA/NCS integration, various processors and various apps.

This PR is the same as #129, minus a handful of the last commits, which
introduce new, unrelated features that should not be part of the same PR.

+++ src/model/caffe_model.cpp
@@ -24,7 +24,7 @@ void CaffeModel<DType>::Load() {
     std::vector<int> gpus;
     GetCUDAGpus(gpus);
 
-    if (desired_device_number < gpus.size()) {
+    if (desired_device_number < (int)gpus.size()) {
       // Device exists
       LOG(INFO) << "Use GPU with device ID " << desired_device_number;
       caffe::Caffe::SetDevice(desired_device_number);
